Refactor home page action code - keypress handler now specified as part of TerminalViewModel
