+++ src/types.ts
@@ -138,9 +138,18 @@ export type Animation =
       };
     };
 
+export interface KeyPressInfo {
+  sequence: string;
+  name: string;
+  ctrl: boolean;
+  meta: boolean;
+  shift: boolean;
+}
+
 export interface TerminalViewModel {
   textWithCursor: TextWithCursor;
   animations: Animation[];
+  keyPressHandler?: (str: string, key: KeyPressInfo) => void;
 }
 
 export type FilesStatus = "user-specified-file-not-found" | "files-found";
