+++ src/view/homePage.ts
@@ -5,6 +5,9 @@ import config from "../config";
 import { HomePage, HomePageData, TerminalViewModel } from "../types";
 import * as renderUtils from "./renderUtils";
 import { version as appVersion } from "../../package.json";
+import * as actions from "../actions";
+import * as debug from "../debug";
+import * as appState from "../appState";
 
 const elideText = (text: string, maxLength: number): string => {
   if (text.length < maxLength) {
@@ -158,5 +161,54 @@ export const render = (
   return {
     textWithCursor: { lines },
     animations: [],
+    keyPressHandler: (str, key) => {
+      debug.log("pressed " + str);
+
+      const state = appState.get();
+      if (state.page.name !== "home") {
+        throw Error("Unexpected page");
+      }
+
+      const fileNameIndex = state.page.selectedFileNameIndex;
+      const topicIndex = state.page.selectedTopicIndex;
+
+      switch (key.name) {
+        case "space":
+        case "return":
+          actions.startSession(homePageData, fileNameIndex, topicIndex);
+          break;
+
+        case "up":
+        case "k":
+          if (topicIndex === 0 && fileNameIndex > 0) {
+            actions.updateHomePage(
+              fileNameIndex - 1,
+              homePageData.topics[fileNameIndex - 1].data.length - 1
+            );
+          } else if (topicIndex > 0) {
+            actions.updateHomePage(fileNameIndex, topicIndex - 1);
+          } else {
+            actions.updateHomePage(fileNameIndex, topicIndex);
+          }
+          break;
+
+        case "down":
+        case "j":
+          if (
+            topicIndex === homePageData.topics[fileNameIndex].data.length - 1 &&
+            fileNameIndex < homePageData.topics.length - 1
+          ) {
+            actions.updateHomePage(fileNameIndex + 1, 0);
+          } else if (
+            topicIndex <
+            homePageData.topics[fileNameIndex].data.length - 1
+          ) {
+            actions.updateHomePage(fileNameIndex, topicIndex + 1);
+          } else {
+            actions.updateHomePage(fileNameIndex, topicIndex);
+          }
+          break;
+      }
+    },
   };
 };
